Update and rename mg_web_node.js to mg_web_node.mjs
